+++ src/collective/es/index/interfaces.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 """Module where all interfaces, events and exceptions live."""
+from . import _
+from zope import schema
 from zope.interface import Interface
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
@@ -20,3 +22,14 @@ class IElasticSearchIndexQueueProcessor(IIndexQueueProcessor):
 
 class IElasticSearchClient(Interface):
     """an initializd  python ElasticSearch object"""
+
+
+class ICollectiveESIndexSettings(Interface):
+
+    index_name = schema.TextLine(
+        title=_(u'Index Name'),
+        description=_(
+            u'help_index_name',
+            default=u'The name of the elasticsearch index used for the site.'
+        ),
+    )